Added comments and TODOs to work through
